refactor: SectionsTest 테스트할 메소드 단위로 파일 분리
